+++ server/config/game/player.js
@@ -10,22 +10,22 @@ const RULES = {
     resourceTypes: [
         {
             type: "energy",
-            start: 200,
+            start: 1000,
             max: {start: 1000, increase: {steps: 5, by: 200}}
         },
         {
             type: "food",
-            start: 200,
+            start: 500,
             max: {start: 500, increase: {steps: 5, by: 100}}
         },
         {
             type: "minerals",
-            start: 200,
+            start: 1000,
             max: {start: 1000, increase: {steps: 5, by: 200}}
         },
         {
             type: "research",
-            start: 200,
+            start: 500,
             max: {start: 500, increase: {steps: 5, by: 100}}
         }
     ]